выглядит как конец

+++ CarRepairDesktop/Views/Clients/MainPage.xaml.cs
@@ -37,7 +37,8 @@ namespace CarRepairDesktop.Views.Clients
 
         private void btnDelete_Click(object sender, RoutedEventArgs e)
         {
-            MessageBox.Show(context.Delete());
+            if (MessageBox.Show("Вы уверены, что хотите удалить запись?", "Удаление", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+                MessageBox.Show(context.Delete());
         }
 
         private void btnEdit_Click(object sender, RoutedEventArgs e)
